// frontend/src/app/shared/generated/models/create-schedule-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateScheduleRequest {
  analysisType?: (string | null);
  assignedToUserId?: (number | string | null);
  exclusionRules?: (string | null);
  name: string;
  recurrencePattern?: (string | null);
  shiftPattern: string;
  site: string;
}