// frontend/src/app/shared/generated/models/update-schedule-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateScheduleRequest {
  analysisType?: (string | null);
  assignedToUserId?: (number | string | null);
  exclusionRules?: (string | null);
  isActive: boolean;
  name: string;
  recurrencePattern?: (string | null);
  rowVersion: string;
  shiftPattern: string;
}